+++ src/app/app.module.ts
@@ -7,6 +7,9 @@ import { FirstComponent } from './first/first.component';
 import { ProductComponent } from './products/components/product/product.component';
 import { ProductListComponent } from './products/components/product-list/product-list.component';
 import { CartListComponent } from './cart/components/cart-list/cart-list.component';
+import {CartModule} from './cart/cart.module';
+import {SharedModule} from './shared/shared.module';
+import {HighlightDirective} from './shared/directives/highlight.directive';
 
 @NgModule({
   declarations: [
@@ -18,7 +21,9 @@ import { CartListComponent } from './cart/components/cart-list/cart-list.compone
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule
+    AppRoutingModule,
+    CartModule,
+    SharedModule
   ],
   providers: [],
   bootstrap: [AppComponent]