+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ProductsService} from '../../services/products.service';
 import {ProductModel} from '../../models/product.model';
+import {CartService} from '../../../cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -10,11 +11,16 @@ import {ProductModel} from '../../models/product.model';
 export class ProductListComponent implements OnInit {
   products!: ProductModel[];
 
-  constructor(private productsService: ProductsService) {
+  constructor(private productsService: ProductsService,
+              private cartService: CartService) {
   }
 
   ngOnInit(): void {
     this.products = this.productsService.getProducts();
+  }
+
+  onAddToCart(product: ProductModel): void {
+    this.cartService.addProductToCart(product);
   }
 
 }