+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,22 @@
+import {Directive, ElementRef, HostBinding, HostListener} from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+
+  constructor(private el: ElementRef) { }
+
+  @HostListener('mouseenter') onMouseEnter(): void {
+    this.highlight('yellow');
+  }
+
+  @HostListener('mouseleave') onMouseLeave(): void {
+    this.highlight('');
+  }
+
+  private highlight(color: string): void {
+    this.el.nativeElement.style.backgroundColor = color;
+  }
+
+}