+++ changelog.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2021-05-30
+### Added
+- Task 2. Components
+- Modules for features (cart, orders, products, shared)
+- Disable 'Add to cart' button if product is not available
+- Show total sum and quantity of products in cart
+- cart-item component
+- Ability to change product quantity in cart
+- Dynamic title in app component
+- Highlight directive
+
 ## [0.0.1] - 2021-05-17
 ### Added
 - Task 1. Introduction.